holds subpages and respective css files
This is implemented with individual CSS files to simplify the creative process for new developers in the group by separating them from the complexities of the website as a whole as much as possible (let them experiment with whatever they want without affecting other pages until group consensus is reached). Individual CSS files will be phased out in later stages.